# docs/swagger.yaml
definitions:
  entity.Artikel:
    properties:
      gambar:
        type: string
      id:
        type: integer
      isi:
        type: string
      jenis:
        type: string
      judul:
        type: string
    type: object
  entity.IMT:
    properties:
      id:
        type: integer
      nama:
        type: string
    type: object
  entity.JadwalMakanan:
    properties:
      id:
        type: integer
      id_resep:
        type: integer
      tanggal:
        type: string
      waktu:
        type: string
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: This is Fits Health API
  title: Fits Health API Documentation
  version: "2.0"
paths:
  /artikel:
    get:
      consumes:
      - application/json
      description: User can Get Information article about health that posting by admin
        or developer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Artikel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Artikel'
      security:
      - JWT: []
      summary: Get All Artikel Information
      tags:
      - Artikel
    post:
      consumes:
      - application/json
      description: Developer can Create Article about healthy, disease, and many more
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.Artikel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Artikel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Artikel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Artikel'
      security:
      - JWT: []
      summary: Create Article
      tags:
      - Artikel
  /artikel/{id}:
    delete:
      consumes:
      - application/json
      description: Developer can Delete Article Information if they want to delete
        it
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Artikel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Artikel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Artikel'
      security:
      - JWT: []
      summary: Delete Artikel
      tags:
      - Artikel
    get:
      consumes:
      - application/json
      description: User can get Article Information by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Artikel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Artikel'
      security:
      - JWT: []
      summary: Get Article Information by ID
      tags:
      - Artikel
    put:
      consumes:
      - application/json
      description: Developer can update about article information that if the information
        article is wrong
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Artikel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Artikel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Artikel'
      security:
      - JWT: []
      summary: Update Artikel
      tags:
      - Artikel
  /imt:
    get:
      consumes:
      - application/json
      description: From IMT information, user can measure itself by knowing what type
        of body he/she get from weight and height
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.IMT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.IMT'
      security:
      - JWT: []
      summary: Get All IMT Information
      tags:
      - IMT
    post:
      consumes:
      - application/json
      description: IMT is for User that categorizing user body by weight and height
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.IMT'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.IMT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.IMT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.IMT'
      security:
      - JWT: []
      summary: Create IMT for User
      tags:
      - IMT
  /imt/{id}:
    delete:
      consumes:
      - application/json
      description: Developer can Delete IMT Information if IMT information is wrong
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.IMT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.IMT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.IMT'
      security:
      - JWT: []
      summary: Delete IMT
      tags:
      - IMT
    get:
      consumes:
      - application/json
      description: User can get IMT Information by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.IMT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.IMT'
      security:
      - JWT: []
      summary: Get IMT Information by ID
      tags:
      - IMT
    put:
      consumes:
      - application/json
      description: Developer can update about IMT information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.IMT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.IMT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.IMT'
      security:
      - JWT: []
      summary: Update IMT
      tags:
      - IMT
  /jadwal-makanan:
    get:
      consumes:
      - application/json
      description: User can get their food schedule information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
      security:
      - JWT: []
      summary: Get All Food Schedule Information in User
      tags:
      - JadwalMakanan
    post:
      consumes:
      - application/json
      description: User can create Food Schedule everyday in their life
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.JadwalMakanan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
      security:
      - JWT: []
      summary: Create Food Schedule
      tags:
      - JadwalMakanan
  /jadwal-makanan/{id}:
    delete:
      consumes:
      - application/json
      description: Developer can Delete Food Schedule Information if they don't like
        about their food schedule that they created
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
      security:
      - JWT: []
      summary: Delete Food Schedule
      tags:
      - JadwalMakanan
    get:
      consumes:
      - application/json
      description: User can get Food Schedule Information by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
      security:
      - JWT: []
      summary: Get Food Schedule Information by ID
      tags:
      - JadwalMakanan
    put:
      consumes:
      - application/json
      description: User can update about Food Schedule if their schdule got wrong
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.JadwalMakanan'
      security:
      - JWT: []
      summary: Update Food Schedule
      tags:
      - JadwalMakanan
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"